event: avoid races between timeout and completion
We need to track the reason passed to the callback in order to
avoid calling the same callback twice in case of races. Therefore,
store the reason for the callback call in the "reason" field of
struct event. As we track the reason for the call in the event
itself now, no need for the extra "reason" argument in callbacks.

Avoid calling timeout and completion for the same event
in a single event_wait() invocation. This is done by calling
the timeout_event last (if it's active), and marking those
events that have completed by setting there reason code.

It's still possible that an event completes between return
from epoll_pwait() and the invocation of the timeout callback.
Therefore callbacks need to check for completion and possibly
act on the event (reset event reason for level-triggered events),
even if called with REASON_TIMEOUT.