timeout: fix moving timeout in corner case
The "position hasn't changed" logic was broken. The unchanged time
stamp case had already been treated. So the only case where nothing
happens is if pnew == pos or pnew == pos + 1 (in both cases the timeout
had changed, but is still larger than the previous and smaller than
the following one, so that the position remains).