event: add cleanup functionality
Add a callback invocation to clean up resources used by the event.
Use a separate callback function here; this avoids having to check
the reason on every callback invocation (cleanup is a rare event).

This is a lot of code for something apparently minor, but
it's important to be able to cleanup the dispatcher structures
and the resources associated with each event e.g. if the main
process forks. Without it, the main program needs to keep track
of every resource used.